pyproject.toml
[tool.poetry]
name = "mysql-visual-explain-cli"
version = "0.1.0"
description = ""
authors = ["Leon Yu <leon@leonyu.net>"]
license = "GPL-2.0-only"
readme = "README.md"
packages = [{include = "mysql_visual_explain_cli"}]

[tool.poetry.dependencies]
python = "^3.11"
cairocffi = "^1.6.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pillow = "^10.4.0"
cairosvg = "^2.7.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
